# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CAIGE - Sistema de Gestão de Idosos API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface LoginRequest {
  email: string;
  senha: string;
}

export interface AuthUser {
  id: number;
  nome: string;
  email: string;
  iniciais: string;
}

export type IdosoSexo = (typeof IdosoSexo)[keyof typeof IdosoSexo];

export const IdosoSexo = {
  masculino: "masculino",
  feminino: "feminino",
} as const;

export type IdosoStatus = (typeof IdosoStatus)[keyof typeof IdosoStatus];

export const IdosoStatus = {
  ativo: "ativo",
  inativo: "inativo",
} as const;

export interface Idoso {
  id: number;
  nome: string;
  idade: number;
  dataNascimento: string;
  sexo: IdosoSexo;
  cpf: string;
  telefone: string;
  bairro: string;
  rua: string;
  numero: string;
  cidade: string;
  estado: string;
  cep: string;
  status: IdosoStatus;
  responsavelNome: string;
  responsavelTelefone: string;
  responsavelParentesco: string;
  observacoes?: string | null;
  cadastro: string;
  iniciais: string;
}

export type CreateIdosoRequestSexo =
  (typeof CreateIdosoRequestSexo)[keyof typeof CreateIdosoRequestSexo];

export const CreateIdosoRequestSexo = {
  masculino: "masculino",
  feminino: "feminino",
} as const;

export type CreateIdosoRequestStatus =
  (typeof CreateIdosoRequestStatus)[keyof typeof CreateIdosoRequestStatus];

export const CreateIdosoRequestStatus = {
  ativo: "ativo",
  inativo: "inativo",
} as const;

export interface CreateIdosoRequest {
  nome: string;
  dataNascimento: string;
  sexo: CreateIdosoRequestSexo;
  cpf: string;
  telefone: string;
  bairro: string;
  rua: string;
  numero: string;
  cidade: string;
  estado: string;
  cep: string;
  status: CreateIdosoRequestStatus;
  responsavelNome: string;
  responsavelTelefone: string;
  responsavelParentesco: string;
  observacoes?: string | null;
}

export interface Prontuario {
  id: number;
  idosoId: number;
  idosoNome: string;
  tipo: string;
  descricao: string;
  responsavel: string;
  data: string;
  hora: string;
  observacoes?: string | null;
}

export interface CreateProntuarioRequest {
  idosoId: number;
  tipo: string;
  descricao: string;
  responsavel: string;
  data: string;
  hora: string;
  observacoes?: string | null;
}

export interface DashboardStats {
  totalIdosos: number;
  profissionaisEnvolvidos: number;
  atendimentosEmAndamento: number;
  atendimentosHoje: number;
  atendimentosSemana: number;
  atendimentosMes: number;
  satisfacao: number;
}

export interface Activity {
  id: number;
  data: string;
  hora: string;
  tipo: string;
  descricao: string;
  responsavel: string;
}

export type ListIdososParams = {
  search?: string;
  status?: ListIdososStatus;
};

export type ListIdososStatus =
  (typeof ListIdososStatus)[keyof typeof ListIdososStatus];

export const ListIdososStatus = {
  ativo: "ativo",
  inativo: "inativo",
} as const;

export type ListProntuariosParams = {
  idosoId?: number;
};