# lib/api-zod/src/generated/types/createProntuarioRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CAIGE - Sistema de Gestão de Idosos API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProntuarioRequest {
  idosoId: number;
  tipo: string;
  descricao: string;
  responsavel: string;
  data: string;
  hora: string;
  observacoes?: string | null;
}