# lib/api-zod/src/generated/types/idoso.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CAIGE - Sistema de Gestão de Idosos API
 * OpenAPI spec version: 0.1.0
 */
import type { IdosoSexo } from "./idosoSexo";
import type { IdosoStatus } from "./idosoStatus";

export interface Idoso {
  id: number;
  nome: string;
  idade: number;
  dataNascimento: string;
  sexo: IdosoSexo;
  cpf: string;
  telefone: string;
  bairro: string;
  rua: string;
  numero: string;
  cidade: string;
  estado: string;
  cep: string;
  status: IdosoStatus;
  responsavelNome: string;
  responsavelTelefone: string;
  responsavelParentesco: string;
  observacoes?: string | null;
  cadastro: string;
  iniciais: string;
}