# lib/api-zod/src/generated/types/loginRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CAIGE - Sistema de Gestão de Idosos API
 * OpenAPI spec version: 0.1.0
 */

export interface LoginRequest {
  email: string;
  senha: string;
}